cleaned up api refactor
