+++ metta/sim/simulation.py
@@ -26,7 +26,7 @@ from omegaconf import OmegaConf
 from metta.agent.policy_record import PolicyRecord
 from metta.agent.policy_state import PolicyState
 from metta.agent.policy_store import PolicyStore
-from metta.api import PreBuiltConfigCurriculum
+from metta.api.environment import PreBuiltConfigCurriculum
 from metta.app_backend.stats_client import StatsClient
 from metta.mettagrid.curriculum.sampling import SamplingCurriculum
 from metta.mettagrid.mettagrid_env import MettaGridEnv, dtype_actions
